+++ src/QuickMarcEditor/constants.js
@@ -6,3 +6,5 @@ export const QUICK_MARC_ACTIONS = {
 };
 
 export const FIELD_TAGS_TO_REMOVE = ['001', '005', '019', '035', '999'];
+
+export const QM_RECORD_STATUS_TIMEOUT = 10000;